requirements.txt
google-auth==2.50.0
google-auth-oauthlib==1.3.1
google-api-python-client==2.195.0
anthropic==0.98.1
python-dotenv==1.2.2
speechrecognition==3.16.1
pyaudio~=0.2.14
openai-whisper==20250625
sounddevice==0.5.5
soundfile==0.13.1
numpy==2.4.4
pytest==9.0.2
tzlocal==5.3.1
requests==2.32.3
spotipy==2.24.0
elevenlabs>=1.0.0
openwakeword
